+++ proto.h
@@ -66,7 +66,8 @@ int get_timestep(int p, double *a, int flag);
 int get_timestep_bin(int ti_step);
 
 double get_starformation_rate(int i);
-void gravity_tree(void);
+void grav_short_tree(void);
+void grav_short_pair(void);
 void hydro_force(void);
 void init(int RestartSnapNum);
 void init_clouds(void);
